HOTT_Service/requirements.txt
Django==1.10.5
argparse==1.2.1
django-crispy-forms==1.6.1
django-filter==1.0.4
django-guardian==1.4.8
djangorestframework==3.6.3
six==1.10.0
wsgiref==0.1.2
